- display search data to ui - make refactor to widget tree
